src/kirbuk_agent/requirements.txt
strands-agents
bedrock-agentcore
strands-agents-tools
playwright
nest-asyncio
sentry-sdk
boto3
